requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
iso8601==0.1.12
Logbook==1.4.0
pyinsales==0.7.5
requests==2.19.1
urllib3==1.23
